PA-2679: Fix backend erroneously rounding values

+++ backend/api/Models/Parcel/ParcelEvaluationModel.cs
@@ -11,7 +11,7 @@ namespace Pims.Api.Models.Parcel
 
         public string Key { get; set; }
 
-        public float Value { get; set; }
+        public decimal Value { get; set; }
 
         public string Note { get; set; }
